CodeJavaApp/src/main/java/com/example/demo/course/CourseService.java
package com.example.demo.course;

import java.util.List;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import com.example.demo.student.Students;

@Service
@Transactional
public class CourseService {
	
    @Autowired
    private CoursesRepository repo;

    public Course get(long id) {
        return repo.findById(id).get();
    }
     
    public void delete(long id) {
        repo.deleteById(id);
    }
}

